test constrain_eval over

+++ src/utils/attribute.h
@@ -3,7 +3,7 @@
 #include <stdbool.h>
 #include <stdint.h>
 
-typedef uint32_t attribute;
+typedef uint64_t attribute;
 
 // Get xn value (0 or 1), USING 1-INDEXING
 bool get_xn(attribute x, int n);
